backend/pyproject.toml
[project]
name = "odoo-backend"
version = "0.1.0"
description = "FastAPI Backend for Odoo Integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.2",
    "python-multipart>=0.0.12",
    "email-validator>=2.2.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "cryptography>=43.0.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "requests-mock>=1.12.1",
    "httpx-mock>=0.0.0"
]
